feat: Add support for resizing images in optimize_images.py
The code changes in optimize_images.py introduce support for resizing images using the scikit-image library. This allows users to specify the width and height of the output images, which will be resized accordingly. The imageio library is also updated to the latest version for image processing.

Based on the recent user commits and repository history, it seems that there have been updates to the optimize_images.py script to improve image processing capabilities. The optimizer.rb file has also been updated to use a relative path for the optimize_images.py script.

Considering the established commit message conventions, the suggested commit message for these code changes is "feat: Add support for resizing images in optimize_images.py". This clearly describes the purpose of the changes and follows the format of a feature-related commit.